requirements.txt
openai==1.14.2
tqdm==4.66.1
requests==2.31.0
prettytable==3.9.0
termcolor==2.4.0
pptree==3.1
climage==0.2.0
PyYAML==6.0.2
tenacity==9.0.0
